+++ packages/common/web/src/Utils/jsonUtils.ts
@@ -1,0 +1,8 @@
+export const safeJsonParse = (jsonString: string) => {
+  try {
+    const parsedJson = JSON.parse(jsonString);
+    return parsedJson;
+  } catch (e) {
+    return undefined;
+  }
+};