+++ packages/plugin-text-color/web/src/index.ts
@@ -3,6 +3,4 @@ export { createTextHighlightPlugin } from './createTextHighlightPlugin';
 export { TEXT_COLOR_TYPE, TEXT_HIGHLIGHT_TYPE } from './types';
 export { Modals, ModalsMap } from './modals';
 export { getColor, isTextColor, isTextHighlight } from './text-decorations-utils';
-export { textColorInlineStyleMapper } from './textColorInlineStyleMapper';
-export { textHighlightInlineStyleMapper } from './textHighlightInlineStyleMapper';
 export { pluginTextColor, pluginTextHighlight } from './editor';